% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caratinga.R
\name{Caratinga}
\alias{Caratinga}
\title{Function El30m - Caratinga (MG)}
\usage{
Caratinga(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Caratinga (MG) (Silva et al., 2010).
}
\details{
Reference: SILVA, M. A. D.; SILVA, M. L. N.; CURI, N.; SANTOS, G. R. D.;
MARQUES, J. J. G. D. S.; MENEZES, M. D. D.; LEITE, F. P. Avaliação e espacialização
da erosividade da chuva no Vale do Rio Doce, região Centro-Leste do Estado de
Minas Gerais. Revista brasileira de ciência do solo, v. 34, p. 1029-1039, 2010.
}
\examples{
Caratinga(RC)


}
