% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dourados.R
\name{Dourados}
\alias{Dourados}
\title{Function El30m - Dourados (MS)}
\usage{
Dourados(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Dourados (MS) (Oliveira et al., 2012).
}
\details{
Reference: OLIVEIRA, P. T.; RODRIGUES, D. B.; SOBRINHO, T. A.;
CARVALHO, D. F. D.; PANACHUKI, E. Spatial variability of the rainfall erosive
potential in the state of Mato Grosso do Sul, Brazil. Engenharia Agrícola,
v. 32, p. 69-79, 2012. DOI: 10.1590/S0100-69162012000100008.
}
\examples{
Dourados(RC)


}
