% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosivity-package.R
\docType{package}
\name{erosivity}
\alias{erosivity}
\alias{_PACKAGE}
\alias{erosivity-package}
\alias{erosivity-package}
\title{erosivity: Compute rainfall erosivity using precipitation data.}
\description{
This package computes rainfall erosivity from the Revised Universal Soil Loss
Equation (RUSLE), using precipitation time series.
}
\section{Data Sources}{


The data are retrieved from:
 \url{http://www.hydroscope.gr/}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/erosivity}
  \item Report bugs at \url{https://github.com/kvantas/erosivity/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
