% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goiania.R
\name{Goiania}
\alias{Goiania}
\title{Function El30m - Goiânia (GO)}
\usage{
Goiania(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 year-1),
according to the calibrated equation for Goiânia (GO) (Silva et al., 1997).
}
\details{
Reference: SILVA, M. L. N.; FREITAS, P. L. de; BLANCANEAUX, P.; CURI, N. Índices
de erosividade das chuvas da região de Goiânia, GO. Pesquisa Agropecuária
Brasileira, v. 32, n. 10, p. 977-985, 1997.
}
\examples{
Goiania(RC)


}
