% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_aggregate.R
\name{hyet_aggregate}
\alias{hyet_aggregate}
\title{Aggregate a hyetograph using a time step}
\usage{
hyet_aggregate(hyet, time_step, units = "mins")
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}

\item{time_step}{an integer specifying the time step value.}

\item{units}{a character string specifying a time unit. Valid units are
"mins" and "hours"}
}
\value{
a tibble with the aggregated hyetograph
}
\description{
\code{hyet_aggregate} uses a predefined time step value to
aggregate precipitation records in a hyetograph
}
\examples{

prec30min <- prec5min \%>\% hyet_aggregate(30)

}
