% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_create.R
\name{hyet_create}
\alias{hyet_create}
\title{Create a hyetograph}
\usage{
hyet_create(prec_dates, prec_values)
}
\arguments{
\item{prec_dates}{a numeric vector of precipitation values}

\item{prec_values}{a POSIXct vector with dates values}
}
\value{
a tibble
}
\description{
\code{hyet_create} uses precipitation and dates values to create a tiblle
that represents the distribution of rainfall over time. This function checks
the validity of precipitation and date values and returns an error if: a)
they don't have the same length, b) precipitation is not numeric or c) dates
is not POSIXct.
}
\examples{

# create dates and precipitation values
prec_dates <- seq(from = as.POSIXct(0, origin = "2018-01-01"),
                  length.out =  100,
                  by = "mins")
set.seed(1)
prec_values <-round(runif(100,0,10),1)

# create hyetograph
hyet <- hyet_create(prec_dates, prec_values)

}
