% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_erosivity.R
\name{hyet_erosivity}
\alias{hyet_erosivity}
\title{Compute raifall erosivity}
\usage{
hyet_erosivity(hyet, time_step)
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}

\item{time_step}{hyetograph's time-step, integer}
}
\value{
a tibble with erosive values of erosive events
}
\description{
\code{prec_eros} computes erosivity values using an hyetograph
from the \code{hyet_create} function and its timestep.
}
\note{
\code{hyet} must not contain missing dates. Please use the
\code{hyet_fill} function before \code{prec_eros}.
}
\examples{

# load data
data(prec5min)

# remove near zero values
prec5min$prec <- round(prec5min$prec , 1)

# compute EI
time_step <- 5
ei_values <- prec5min \%>\%
 hyet_fill(time_step) \%>\%
 hyet_erosivity(time_step)

}
