% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_fill.R
\name{hyet_fill}
\alias{hyet_fill}
\title{Fill a hyetograph with missing dates values}
\usage{
hyet_fill(hyet, time_step)
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}

\item{time_step}{hyetograph's time-step}
}
\value{
a tibble with the variables \code{dates} and \code{prec}
}
\description{
\code{hyet_fill} fills missing dates values in a hyetograph
with NA values.  Returns an error if \code{hyet} is not a valid hyetograph or
\code{time_step} is a not a numeric value.
}
\examples{

# create dates and precipitation values using 5 minutes time-step
prec_dates <- seq(from = as.POSIXct(0, origin = "2018-01-01"),
                  length.out =  100,
                  by = "5 mins")
set.seed(1)
prec_values <-round(runif(100,0,10),1)

# create hyetograph
hyet <- hyet_create(prec_dates, prec_values)

# remove some random values from hyetograph
hyet_miss <- hyet[-sample(100,30), ]

# fill hyetograph
hyet_fill(hyet_miss, 5)
}
