% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_remove_rep.R
\name{hyet_remove_rep}
\alias{hyet_remove_rep}
\title{Replace repeated, non-zero, precipitation values with NA}
\usage{
hyet_remove_rep(hyet)
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}
}
\value{
a tibble with the variables \code{dates} and \code{prec}
}
\description{
\code{hyet_remove_rep} sets to \code{NA} repeated values
precipitation values in a hyetograph. Returns an error if \code{hyet} is not
a valid hyetograph.
}
\examples{

# create a hyetograph
dates <- seq(from = as.POSIXct(0, origin = "2018-01-01"),
                  length.out =  11,
                  by = "30 mins")
prec <- c(0.001, 0.2, 0.5, 0.21, 0.21, 0.21, 0.21, 0.21, 0.0005, 0, 0)
hyet <- hyet_create(dates, prec)

# remove repeated values
hyet_remove_rep(hyet)

}
