% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_window_sum.R
\name{hyet_window_sum}
\alias{hyet_window_sum}
\title{Calculate rolling window sum}
\usage{
hyet_window_sum(prec, rolling_window)
}
\arguments{
\item{prec}{a vector with precipitation values}

\item{rolling_window}{the rolling window as a number of time steps.

Computing time depends on \code{rolling_window}. \code{hyet_window_sum}
use R's base \code{rowSums} function with \code{na.rm = TRUE}.}
}
\value{
a numeric vector with summed precipitation values. These values are
filled with NAs at the end.
}
\description{
\code{hyet_window_sum} computes the rolling sum of precipitation
values.  Returns an error if \code{prec} is not a numeric vector or
\code{rolling_window} not a positive integer.
}
\examples{

# load data
data(prec5min)

# fill hyetograph
hyet <- hyet_fill(prec5min, 5)

# set rolling window sum to 6 values
rolling_window <- 6

# compute values
hyet_window_sum(hyet$prec, rolling_window)
}
