% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lavras.R
\name{Lavras}
\alias{Lavras}
\title{Function El30m - Lavras (MG)}
\usage{
Lavras(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Lavras (MG) (Aquino et al., 2014).
}
\details{
Reference: AQUINO, R. F.; SILVA, M. L. N.; FREITAS, D. A. F. D.; CURI, N.;
MELLO, C. R. D.; AVANZI, J. C. Erosividade das chuvas e tempo de recorrência
para Lavras, Minas Gerais. Revista Ceres, v. 61, p. 09-16, 2014.
}
\examples{
Lavras(RC)


}
