% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poxoreu.R
\name{Poxoreu}
\alias{Poxoreu}
\title{Function El30m - Poxoréu (MT)}
\usage{
Poxoreu(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Poxoréu (MT) (Di Raimo et al., 2018).
}
\details{
Reference: DI RAIMO, L. A. DI L.; AMORIM, R. S. S.; COUTO, E. G.; NÓBREGA,
R. L. B.; TORRES, G. N.; BOCUTI, E. D.; ALMEIDA, C. O. S.; RODRIGUES,
R. V. Spatio-temporal variability of erosivity in Mato Grosso. Brazil.
Revista Ambiente e Água, v. 13. n. 6. p. 1- 14. 2018. DOI: 10.4136/1980-993X.
}
\examples{
Poxoreu(RC)


}
