% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prec10min}
\alias{prec10min}
\title{Ten minutes precipitation time series}
\format{A data frame with 52,116 rows and 2 variables:
\describe{
  \item{date}{The time series' dates}
  \item{prec}{precipitation, in mm}
}}
\source{
Meusburger, K., Steel, A., Panagos, P., Montanarella, L., Alewell, C.
Spatial and temporal variability of rainfall erosivity factor for Switzerland.
 Hydrology and Earth System Sciences, 16, 167-177, 2012, URL:
\url{https://esdac.jrc.ec.europa.eu/content/spatial-and-temporal-variability-rainfall-erosivity-factor-switzerland}
}
\usage{
prec10min
}
\description{
The time series come from a weather station in Switzerland. The units of
measurement of precipitation is in mm and the time step is five minutes.
}
\keyword{datasets}
