% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prec5min}
\alias{prec5min}
\title{Five minutes precipitation time series}
\format{A data frame with 48,209 rows and 2 variables:
\describe{
  \item{date}{The time series' dates}
  \item{prec}{precipitation, in mm}
}}
\source{
\url{http://kyy.hydroscope.gr/timeseries/d/1432/}
}
\usage{
prec5min
}
\description{
This time series come from the a wether station from the Greek National Data
Bank for Hydrological and Meteorological Information, Hydroscope.
Precipitation's units of measurement is in mm and the time step is five
minutes.
}
\keyword{datasets}
