% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{RC}
\alias{RC}
\title{Function for the rainfall coefficient}
\usage{
RC(MM, MA)
}
\arguments{
\item{MM}{Monthly average precipitation}

\item{MA}{Annual average precipitation}
}
\description{
This function is used to calculate the rainfall coefficient parameter (mm),
proposed by Fournier (1956) and modified by Lombardi Neto (1977).
}
\details{
Reference: LOMBARDI NETO, F. Rainfall erosivity - its distribution and
relationship with soil loss at Campinas, Brasil. West Lafayette,
Purdue University, 1977. 53p. (Masters dissertation).
}
\examples{
RC(230, 1300)


}
