% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regiao_1.R
\name{Regiao_1}
\alias{Regiao_1}
\title{Function El30m  - Região 1 (PR)}
\usage{
Regiao_1(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Região 1 in Paraná (Rufino et al., 1993).
}
\details{
Reference: Rufino RL, Biscaia RCM, Merten GH. Determinação do potencial
erosivo da chuva do estado do Paraná através da pluviometria: Terceira
aproximação. R Bras Ci Solo. 1993;17:439-44.
}
\examples{
Regiao_1(RC)


}
