% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sete_Lagoas.R
\name{Sete_Lagoas}
\alias{Sete_Lagoas}
\title{Function El30m - Sete Lagoas (MG)}
\usage{
Sete_Lagoas(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Sete Lagoas (MG) (Marques et al., 1998).
}
\details{
Reference: MARQUES, M. DE SÁ J. J. G.; ALVARENGA, R. C.; CURI, N. Erosividade
das chuvas da região de Sete Lagoas, MG. Pesquisa Agropecuária Brasileira,
v. 33, n. 5, p. 761-768, 1998.
}
\examples{
Sete_Lagoas(RC)


}
