% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Teodoro_Sampaio.R
\name{Teodoro_Sampaio}
\alias{Teodoro_Sampaio}
\title{Function El30m - Teodoro e Sampaio (SP)}
\usage{
Teodoro_Sampaio(RC)
}
\arguments{
\item{RC}{rainfall coefficient}
}
\description{
This is the function to calculate monthly erosivity (MJ mm ha-1 h-1 ano-1),
according to the calibrated equation for Teodoro e Sampaio (SP) (Colodro et al., 2002).
}
\details{
Reference: COLODRO, G.; CARVALHO, M. P.; ROQUE, C. G.; PRADO, R. M.
Erosividade da chuva: distribuição e correlação com a precipitação pluviométrica
de Teodoro Sampaio (SP). Revista Brasileira de Ciência do Solo, v. 26, p. 809-818, 2002.
}
\examples{
Teodoro_Sampaio(RC)


}
