% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_aggregate.R
\name{hyet_aggregate}
\alias{hyet_aggregate}
\title{Aggregate a hyetofraph using a time step}
\usage{
hyet_aggregate(hyet, time_step_minutes)
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}

\item{time_step_minutes}{the time step value in minutes}
}
\value{
a tibble with the aggregated hyetograph
}
\description{
\code{hyet_aggregate} uses a predefined time step value to
aggregate precipitation records in a hyetograph
}
\examples{

prec30min <- prec5min \%>\% hyet_aggregate(30)

}
