% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prec5min}
\alias{prec5min}
\title{Five minutes precipitation time series}
\format{A data frame with 1679308 rows and 13 variables:
\describe{
  \item{date}{The time series' date}
  \item{value}{precipitation, in mm}
}}
\source{
\url{http://kyy.hydroscope.gr/timeseries/d/1432/}
}
\usage{
prec5min
}
\description{
The time series come from the Greek National Data Bank for Hydrological
and Meteorological Information, Hydroscope. The unit of measurement of
precipitation is in mm and the time scale is five minutes.
}
\keyword{datasets}
